refactor(flow-info): Move flow-info to its own crate
In prevision for more integration with the dataplane (in particular,
starting with the NAT code), move the flow-info code out of pkt-meta and
into its own crate, with the objective to avoid the risk of circular
dependencies.

Suggested-by: Manish Vachharajani <manish@githedgehog.com>
