+++ flow-info/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "dataplane-flow-info"
+version = "0.1.0"
+edition = "2024"
+publish = false
+license = "Apache-2.0"
+
+[dependencies]
+atomic-instant-full = { workspace = true }
+concurrency = { workspace = true }
+net = { workspace = true }
+thiserror = { workspace = true }
