+++ pkt-meta/src/flow_table/mod.rs
@@ -1,17 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Open Network Fabric Authors
 
-pub mod atomic_instant;
-pub mod flow_info;
 pub mod flow_key;
 pub mod table;
 mod thread_local_pq;
 
-pub use flow_info::*;
 pub use flow_key::IpProtoKey;
 pub use flow_key::TcpProtoKey;
 pub use flow_key::UdpProtoKey;
 pub use flow_key::{FlowKey, FlowKeyData};
 pub use table::FlowTable;
-
-pub use atomic_instant::AtomicInstant;
